// src/types/mapbox_style.ts
export type ColorSpecification = string

export type FormattedSpecification = string

export type ResolvedImageSpecification = string

export type PromoteIdSpecification =
  | {
      [_: string]: string
    }
  | string
// ...
export type TransitionSpecification = {
  duration?: number
  delay?: number
}
// ...
export type PropertyValueSpecification<T> =
  | T
  | CameraFunctionSpecification<T>
  | ExpressionSpecification

export type DataDrivenPropertyValueSpecification<T> =
  | T
  | CameraFunctionSpecification<T>
  | SourceFunctionSpecification<T>
  | CompositeFunctionSpecification<T>
  | ExpressionSpecification

export type StyleSpecification = {
  /** URL where this style was downloaded from */
  upstreamUrl?: string
  version: 8
  name?: string
  metadata?: unknown
  center?: Array<number>
  zoom?: number
  bearing?: number
  pitch?: number
  light?: LightSpecification
  sources: {
    [_: string]: SourceSpecification
  }
  sprite?: string
  glyphs?: string
  transition?: TransitionSpecification
  layers: Array<LayerSpecification>
}

export type LightSpecification = {
  anchor?: PropertyValueSpecification<'map' | 'viewport'>
  position?: PropertyValueSpecification<[number, number, number]>
  color?: PropertyValueSpecification<ColorSpecification>
  intensity?: PropertyValueSpecification<number>
}

export type VectorSourceSpecification = {
  type: 'vector'
  url?: string
  tiles?: Array<string>
  bounds?: [number, number, number, number]
  scheme?: 'xyz' | 'tms'
  minzoom?: number
  maxzoom?: number
  attribution?: string
  promoteId?: PromoteIdSpecification
  volatile?: boolean
}

export type RasterSourceSpecification = {
  type: 'raster'
  url?: string
  tiles?: Array<string>
  bounds?: [number, number, number, number]
  minzoom?: number
  maxzoom?: number
  tileSize?: number
  scheme?: 'xyz' | 'tms'
  attribution?: string
  volatile?: boolean
}

export type RasterDEMSourceSpecification = {
  type: 'raster-dem'
  url?: string
  tiles?: Array<string>
  bounds?: [number, number, number, number]
  minzoom?: number
  maxzoom?: number
  tileSize?: number
  attribution?: string
  encoding?: 'terrarium' | 'mapbox'
  volatile?: boolean
}

export type GeoJSONSourceSpecification = {
  type: 'geojson'
  data?: unknown
  maxzoom?: number
  attribution?: string
  buffer?: number
  filter?: unknown
  tolerance?: number
  cluster?: boolean
  clusterRadius?: number
  clusterMaxZoom?: number
  clusterMinPoints?: number
  clusterProperties?: unknown
  lineMetrics?: boolean
  generateId?: boolean
  promoteId?: PromoteIdSpecification
}

export type VideoSourceSpecification = {
  type: 'video'
  urls: Array<string>
  coordinates: [
    [number, number],
    [number, number],
    [number, number],
    [number, number]
  ]
}

export type ImageSourceSpecification = {
  type: 'image'
  url: string
  coordinates: [
    [number, number],
    [number, number],
    [number, number],
    [number, number]
  ]
}

export type SourceSpecification =
  | VectorSourceSpecification
  | RasterSourceSpecification
  | RasterDEMSourceSpecification
  | GeoJSONSourceSpecification
  | VideoSourceSpecification
  | ImageSourceSpecification

export type FillLayerSpecification = {
  id: string
  type: 'fill'
  metadata?: unknown
  source: string
  'source-layer'?: string
  minzoom?: number
  maxzoom?: number
  filter?: FilterSpecification
  layout?: {
    'fill-sort-key'?: DataDrivenPropertyValueSpecification<number>
    visibility?: 'visible' | 'none'
  }
  paint?: {
    'fill-antialias'?: PropertyValueSpecification<boolean>
    'fill-opacity'?: DataDrivenPropertyValueSpecification<number>
    'fill-color'?: DataDrivenPropertyValueSpecification<ColorSpecification>
    'fill-outline-color'?: DataDrivenPropertyValueSpecification<
      ColorSpecification
    >
    'fill-translate'?: PropertyValueSpecification<[number, number]>
    'fill-translate-anchor'?: PropertyValueSpecification<'map' | 'viewport'>
    'fill-pattern'?: DataDrivenPropertyValueSpecification<
      ResolvedImageSpecification
    >
  }
}

export type LineLayerSpecification = {
  id: string
  type: 'line'
  metadata?: unknown
  source: string
  'source-layer'?: string
  minzoom?: number
  maxzoom?: number
  filter?: FilterSpecification
  layout?: {
    'line-cap'?: PropertyValueSpecification<'butt' | 'round' | 'square'>
    'line-join'?: DataDrivenPropertyValueSpecification<
      'bevel' | 'round' | 'miter'
    >
    'line-miter-limit'?: PropertyValueSpecification<number>
    'line-round-limit'?: PropertyValueSpecification<number>
    'line-sort-key'?: DataDrivenPropertyValueSpecification<number>
    visibility?: 'visible' | 'none'
  }
  paint?: {
    'line-opacity'?: DataDrivenPropertyValueSpecification<number>
    'line-color'?: DataDrivenPropertyValueSpecification<ColorSpecification>
    'line-translate'?: PropertyValueSpecification<[number, number]>
    'line-translate-anchor'?: PropertyValueSpecification<'map' | 'viewport'>
    'line-width'?: DataDrivenPropertyValueSpecification<number>
    'line-gap-width'?: DataDrivenPropertyValueSpecification<number>
    'line-offset'?: DataDrivenPropertyValueSpecification<number>
    'line-blur'?: DataDrivenPropertyValueSpecification<number>
    'line-dasharray'?: PropertyValueSpecification<Array<number>>
    'line-pattern'?: DataDrivenPropertyValueSpecification<
      ResolvedImageSpecification
    >
    'line-gradient'?: ExpressionSpecification
  }
}

export type SymbolLayerSpecification = {
  id: string
  type: 'symbol'
  metadata?: unknown
  source: string
  'source-layer'?: string
  minzoom?: number
  maxzoom?: number
  filter?: FilterSpecification
  layout?: {
    'symbol-placement'?: PropertyValueSpecification<
      'point' | 'line' | 'line-center'
    >
    'symbol-spacing'?: PropertyValueSpecification<number>
    'symbol-avoid-edges'?: PropertyValueSpecification<boolean>
    'symbol-sort-key'?: DataDrivenPropertyValueSpecification<number>
    'symbol-z-order'?: PropertyValueSpecification<
      'auto' | 'viewport-y' | 'source'
    >
    'icon-allow-overlap'?: PropertyValueSpecification<boolean>
    'icon-ignore-placement'?: PropertyValueSpecification<boolean>
    'icon-optional'?: PropertyValueSpecification<boolean>
    'icon-rotation-alignment'?: PropertyValueSpecification<
      'map' | 'viewport' | 'auto'
    >
    'icon-size'?: DataDrivenPropertyValueSpecification<number>
    'icon-text-fit'?: PropertyValueSpecification<
      'none' | 'width' | 'height' | 'both'
    >
    'icon-text-fit-padding'?: PropertyValueSpecification<
      [number, number, number, number]
    >
    'icon-image'?: DataDrivenPropertyValueSpecification<
      ResolvedImageSpecification
    >
    'icon-rotate'?: DataDrivenPropertyValueSpecification<number>
    'icon-padding'?: PropertyValueSpecification<number>
    'icon-keep-upright'?: PropertyValueSpecification<boolean>
    'icon-offset'?: DataDrivenPropertyValueSpecification<[number, number]>
    'icon-anchor'?: DataDrivenPropertyValueSpecification<
      | 'center'
      | 'left'
      | 'right'
      | 'top'
      | 'bottom'
      | 'top-left'
      | 'top-right'
      | 'bottom-left'
      | 'bottom-right'
    >
    'icon-pitch-alignment'?: PropertyValueSpecification<
      'map' | 'viewport' | 'auto'
    >
    'text-pitch-alignment'?: PropertyValueSpecification<
      'map' | 'viewport' | 'auto'
    >
    'text-rotation-alignment'?: PropertyValueSpecification<
      'map' | 'viewport' | 'auto'
    >
    'text-field'?: DataDrivenPropertyValueSpecification<FormattedSpecification>
    'text-font'?: DataDrivenPropertyValueSpecification<Array<string>>
    'text-size'?: DataDrivenPropertyValueSpecification<number>
    'text-max-width'?: DataDrivenPropertyValueSpecification<number>
    'text-line-height'?: PropertyValueSpecification<number>
    'text-letter-spacing'?: DataDrivenPropertyValueSpecification<number>
    'text-justify'?: DataDrivenPropertyValueSpecification<
      'auto' | 'left' | 'center' | 'right'
    >
    'text-radial-offset'?: DataDrivenPropertyValueSpecification<number>
    'text-variable-anchor'?: PropertyValueSpecification<
      Array<
        | 'center'
        | 'left'
        | 'right'
        | 'top'
        | 'bottom'
        | 'top-left'
        | 'top-right'
        | 'bottom-left'
        | 'bottom-right'
      >
    >
    'text-anchor'?: DataDrivenPropertyValueSpecification<
      | 'center'
      | 'left'
      | 'right'
      | 'top'
      | 'bottom'
      | 'top-left'
      | 'top-right'
      | 'bottom-left'
      | 'bottom-right'
    >
    'text-max-angle'?: PropertyValueSpecification<number>
    'text-writing-mode'?: PropertyValueSpecification<
      Array<'horizontal' | 'vertical'>
    >
    'text-rotate'?: DataDrivenPropertyValueSpecification<number>
    'text-padding'?: PropertyValueSpecification<number>
    'text-keep-upright'?: PropertyValueSpecification<boolean>
    'text-transform'?: DataDrivenPropertyValueSpecification<
      'none' | 'uppercase' | 'lowercase'
    >
    'text-offset'?: DataDrivenPropertyValueSpecification<[number, number]>
    'text-allow-overlap'?: PropertyValueSpecification<boolean>
    'text-ignore-placement'?: PropertyValueSpecification<boolean>
    'text-optional'?: PropertyValueSpecification<boolean>
    visibility?: 'visible' | 'none'
  }
  paint?: {
    'icon-opacity'?: DataDrivenPropertyValueSpecification<number>
    'icon-color'?: DataDrivenPropertyValueSpecification<ColorSpecification>
    'icon-halo-color'?: DataDrivenPropertyValueSpecification<ColorSpecification>
    'icon-halo-width'?: DataDrivenPropertyValueSpecification<number>
    'icon-halo-blur'?: DataDrivenPropertyValueSpecification<number>
    'icon-translate'?: PropertyValueSpecification<[number, number]>
    'icon-translate-anchor'?: PropertyValueSpecification<'map' | 'viewport'>
    'text-opacity'?: DataDrivenPropertyValueSpecification<number>
    'text-color'?: DataDrivenPropertyValueSpecification<ColorSpecification>
    'text-halo-color'?: DataDrivenPropertyValueSpecification<ColorSpecification>
    'text-halo-width'?: DataDrivenPropertyValueSpecification<number>
    'text-halo-blur'?: DataDrivenPropertyValueSpecification<number>
    'text-translate'?: PropertyValueSpecification<[number, number]>
    'text-translate-anchor'?: PropertyValueSpecification<'map' | 'viewport'>
  }
}

export type CircleLayerSpecification = {
  id: string
  type: 'circle'
  metadata?: unknown
  source: string
  'source-layer'?: string
  minzoom?: number
  maxzoom?: number
  filter?: FilterSpecification
  layout?: {
    'circle-sort-key'?: DataDrivenPropertyValueSpecification<number>
    visibility?: 'visible' | 'none'
  }
  paint?: {
    'circle-radius'?: DataDrivenPropertyValueSpecification<number>
    'circle-color'?: DataDrivenPropertyValueSpecification<ColorSpecification>
    'circle-blur'?: DataDrivenPropertyValueSpecification<number>
    'circle-opacity'?: DataDrivenPropertyValueSpecification<number>
    'circle-translate'?: PropertyValueSpecification<[number, number]>
    'circle-translate-anchor'?: PropertyValueSpecification<'map' | 'viewport'>
    'circle-pitch-scale'?: PropertyValueSpecification<'map' | 'viewport'>
    'circle-pitch-alignment'?: PropertyValueSpecification<'map' | 'viewport'>
    'circle-stroke-width'?: DataDrivenPropertyValueSpecification<number>
    'circle-stroke-color'?: DataDrivenPropertyValueSpecification<
      ColorSpecification
    >
    'circle-stroke-opacity'?: DataDrivenPropertyValueSpecification<number>
  }
}
// ...
export type BackgroundLayerSpecification = {
  id: string
  type: 'background'
  metadata?: unknown
  minzoom?: number
  maxzoom?: number
  layout?: {
    visibility?: 'visible' | 'none'
  }
  paint?: {
    'background-color'?: PropertyValueSpecification<ColorSpecification>
    'background-pattern'?: PropertyValueSpecification<
      ResolvedImageSpecification
    >
    'background-opacity'?: PropertyValueSpecification<number>
  }
}

export type LayerSpecification =
  | FillLayerSpecification
  | LineLayerSpecification
  | SymbolLayerSpecification
  | CircleLayerSpecification
  | HeatmapLayerSpecification
  | FillExtrusionLayerSpecification
  | RasterLayerSpecification
  | HillshadeLayerSpecification
  | BackgroundLayerSpecification